requirements.txt
aiofiles==24.1.0
blinker==1.9.0
click==8.1.7
exceptiongroup==1.2.2
flask==3.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.17.3
hyperframe==6.0.1
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==3.0.2
priority==2.0.0
python-dotenv==1.0.1
quart==0.19.9
taskgroup==0.0.0a4
tomli==2.2.1
typing-extensions==4.12.2
werkzeug==3.1.3
wsproto==1.2.0
zipp==3.21.0
